Can get todos, Can Update todos

+++ src/todo/todo.service.ts
@@ -2,6 +2,7 @@ import { PrismaService } from "src/prisma/prisma.service";
 import { ForbiddenException, Injectable } from '@nestjs/common';
 import { Prisma, User } from '@prisma/client';
 import { TodoDto } from './dto';
+import { Status } from './enum';
 
 @Injectable()
 export class TodoService {
@@ -14,7 +15,7 @@ export class TodoService {
         data: {
           name: dto.name,
           description: dto.description,
-          status: 'NotStarted',
+          status: dto.status,
           userId: user.id
         }
       })
@@ -24,15 +25,90 @@ export class TodoService {
     }
   }
 
-  async deleteTodo(id: string){
+  async findTodosByName(status: Status, user: User){
+    if(status.length === 0){
+      try {
+        const allTodos = await this.prisma.todo.findMany({
+          where:{
+            userId: user.id
+          }
+        });
 
-    try{
-      const todoId = Number(id);
-      const todo = await this.prisma.todo.delete({
+        return allTodos;
+      } catch(error) {
+        throw error;
+      }
+    }
+    try {
+      const todos = await this.prisma.todo.findMany({
         where:{
-          id: todoId
+          status: status,
+          userId: user.id
         }
       })
+      return todos;
+    } catch(error) {
+      throw error;
+    }
+  }
+
+  async updateTodo(todoId: number, dto: TodoDto, user: User){
+    try{
+      // if right user for todo
+      const todoRightOwner= await this.prisma.todo.findFirst({
+        where:{
+          id: todoId,
+          userId: user.id
+        }
+      })
+      // Deny if owner false
+      if(todoRightOwner == null) throw new ForbiddenException('Permission denied');
+
+      console.log(dto.description);
+
+      const todo = await this.prisma.todo.update({
+        where:{
+          id: todoId
+        },
+        data: {
+          name: dto.name,
+          description: dto.description,
+          status: dto.status,
+        }
+      })
+      return {message: 'Updated : ' + todo.name};
+    } catch(error) {
+      if(error instanceof Prisma.PrismaClientKnownRequestError) {
+        if (error.code == 'P2025') {
+          throw new ForbiddenException('Todo is not in Database');
+        }
+      }
+      throw error;
+    }
+  }
+
+  async deleteTodo(todoId: number, user: User){
+    try{
+      // if right user for todo
+      const todoRightOwner= await this.prisma.todo.findFirst({
+        where:{
+          id: todoId,
+          userId: user.id
+        }
+      })
+      // Deny if owner false
+      if(todoRightOwner == null) throw new ForbiddenException('Permission denied');
+
+      const todo = await this.prisma.todo.delete({
+        where:{
+          id: todoId,
+        },
+        select: {
+          name: true,
+        }
+      })
+
+      return {message: 'removed todo ' + todo.name};
     } catch(error) {
       if(error instanceof Prisma.PrismaClientKnownRequestError) {
         if (error.code == 'P2025') {
@@ -42,6 +118,5 @@ export class TodoService {
       throw error;
     }
 
-    return {message: 'removed ' + id};
   }
 }
